Create a workflow to upload binary releases for windows
Compiling this library from source is not only computationally
expensive, but also requires users to have a functioning Rust
toolchain installed which is a non-starter for many developers.

This uses [`node-pre-gyp`][1] to build the binaries ahead of time on
github actions so that users can downlad the binaries for their
particular platform. For now, we will use node 10,12, and 14

Caveat:
For now, the version of node-pre-gyp which is compatible with neon is
not released to NPM and is only available on github, so the end system
must have GIT installed in order to successfully NPM install which
kinda stinks.

[1]: https://www.npmjs.com/package/node-pre-gyp