Use "killer" strategy
We were having race conditions where the console of the parent process
console was getting into an inconsistent state. We think this was
because we were calling FreeConsole, AttachConsole, FreeConsole, and
AttachConsole again.

This avoids having to make the parent process detach from its console
temporarily by spawning a "killer" process whose only purpose is to
throw away its own console, attach to the target process's console and
then send a ctrl-c event to it. Then the "killer" process can die
itself and we never have to worry about the parent process.

We still do have to temporarily disable ctrl-c handling in the parent
process since it is possible that it shares a console with the target
process.