Add a clean script
A default yarn install will now try to get the binary from github so
we may need to start clean. Also, it's handly when you want to build
completely from scratch.

This delegates cleaning to the neon CLI for the most part, but also
removes our custom `killer.exe` binary