Fix binary releases
The neon docs recommend using some weird fork of node-pre-gyp, but as
we found out to our chagrin, it was generating binary tarbals that
mixed the lib directory with the native directory.

This completely side-steps the issue by copying built artifacts into
the `dist/` directory and then using the normal un-forked node-pre-gyp
npm package.