Always skip build on non-windows platforms.
When installing ctrlc-windows on a non-windows platform, the
`node-pre-gyp` invocation fails because a pre-compiled binary does not
exist. This isn't much of a problem since the fallback behavior is to
run the `build.js` script which skips non-windows platforms
entirely. In other words, the command is roughly

```sh
$ node-pre-gyp || node build.js
```

And it succeeds, so `yarn` does not bother you with the fact that
`node-pre-gyp` failed.

However, when using `npm`, not yarn, the `stderr` is printed
to the console always, even though the package install script succeeds
overally. We want this warning to not be displayed with either package
manager.

This adds a `not-windows.js` script that exits successfully for all
non-windows platforms that can be used to effectively "short-circuit"
the install script if we're not running on windows. Not only does this
silence the warning, but also has the added benefit that
`node-pre-gyp` is never invoked and so the network is never queried
for the binaries that will never be found for non-windows systems.